:bug: hook conflicst with merge resolved incorrectly
