+++ src/components/RulesTable/index.tsx
@@ -1,11 +1,8 @@
 import { FolderOff, ReadMore } from '@mui/icons-material';
 import { CircularProgress, TableCell, TablePagination, TableRow, Typography } from '@mui/material'
-import relative from 'dayjs/plugin/relativeTime'
-import 'dayjs/locale/es'
-import dayjs from 'dayjs';
 
 import GenericTable from '../../layout/GenericTable'
-import { StatusIcon } from '../StatusIcon';
+
 import { GenericPopover } from '../GenericPopover';
 import { UncontrolledSwitch } from '../Switch/uncontrolledIndexed';
 import { columns } from './config';
@@ -14,23 +11,17 @@ import { useGetRulesStatus } from '../../hooks';
 import { GenericHeader } from '../GenericHeader';
 import { Info } from '../../layout/Info';
 import { MouseEvent, useCallback, useState } from 'react';
+import { TimeAgo } from '../TimeAgo';
 
-dayjs.extend(relative)
-dayjs.locale('es')
 
-interface TimeAgoProps {
-  date: string | Date
-}
-
-const TimeAgo = ({date}: TimeAgoProps)=> <Typography>
-    {dayjs(date).from(new Date())}
-  </Typography>
 
 
 // TODO: check another abstraction for especial cases
 const EspecialConfigCell = ({resource, value, id} : {resource: string; value: unknown; id: string;})=>{
   if(resource === 'isActiveSonar')
-    return <StatusIcon isActive={Boolean(value)}/>
+    return <Typography>
+      {value ? 'Activo' : 'No activo'}
+    </Typography>
 
   if (resource === 'htmlDesc')
     return (<GenericPopover
